user controller integration tests working
