+++ chargingplatform/src/test/java/nikev/group/project/chargingplatform/FunctionalTests/Hooks.java
@@ -1,0 +1,34 @@
+package nikev.group.project.chargingplatform.FunctionalTests;
+
+import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.chrome.ChromeDriver;
+import org.openqa.selenium.chrome.ChromeOptions;
+
+import io.cucumber.java.After;
+import io.cucumber.java.Before;
+import io.github.bonigarcia.wdm.WebDriverManager;
+
+public class Hooks {
+    public static WebDriver driver;
+    
+    @Before
+    public void setUp() {
+        WebDriverManager.chromedriver().setup();
+        ChromeOptions options = new ChromeOptions();
+
+        options.addArguments("--no-sandbox");
+        options.addArguments("--disable-dev-shm-usage");
+        options.addArguments("--disable-gpu");
+        options.addArguments("--remote-debugging-port=9222");
+        options.addArguments("--window-size=1920,1080");
+        driver = new ChromeDriver(options);
+    }
+
+    @After
+    public void cleanUp() {
+        if (null != driver) {
+            driver.quit();
+        }
+    }
+
+}
